+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Add support for stacked images, [PR-11](https://github.com/panda-official/DriftCLI/pull/11)
+
 ## 0.5.0 - 2023-04-06
 
 ### Added